+++ packages/koot/core/pwa/create.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 const fs = require('fs-extra');
 const path = require('path');
 const glob = require('glob-promise');
@@ -38,11 +39,9 @@ const parsePattern = pattern => {
  * @param {string} [settings.outputFilename=`service-worker.js`] - 输出文件名
  * @param {boolean} [settings.outputFilenameHash=false] - 输出文件名中是否会加入 MD5 hash。若设为 true，输出文件名会变为 `${basename}.${md5hash}.js`
  * @param {string} [settings.customServiceWorkerPath=path.resolve(__dirname, '../service-worker/index.js')] - service-worker 文件模板
- * @param {string} [settings.globPattern=`/client/**/ const create = async (
-    settings = {},
-    i18n,
-    bundleVersionsKeep
-) => {
+ * @param {string} [settings.globPattern=`/client/`]
+ **/
+const create = async (settings = {}, i18n) => {
     // let options = Object.assign({
     //     outputPath: getCwd() + '/dist/public/',
     //     outputFilename: 'service-worker.js',
@@ -65,7 +64,7 @@ const parsePattern = pattern => {
 
     const pathnamePolyfill = [];
     const pathnameChunkmap = getChunkmapPath();
-    const outputPath = getDirDistPublic(getDistPath(), bundleVersionsKeep);
+    const outputPath = getDirDistPublic(getDistPath());
     const i18nType = typeof i18n === 'object' ? i18n.type : undefined;
     const isI18nDefault = i18nType === 'default';