+++ packages/koot/defaults/before-build.js
@@ -50,6 +50,8 @@ module.exports = {
     WEBPACK_OUTPUT_PATH: '__WEBPACK_OUTPUT_PATH',
     CLIENT_ROOT_PATH: '__CLIENT_ROOT_PATH',
     WEBPACK_MODIFIED_PUBLIC_PATH: '__WEBPACK_MODIFIED_PUBLIC_PATH',
+    // CLEANUP_PREVIOUS_BUILDS_IDS: '__CLEANUP_PREVIOUS_BUILDS_IDS',
+    // CLEANUP_PREVIOUS_BUILDS_FILES: '__CLEANUP_PREVIOUS_BUILDS_FILES',
 
     // 自定义 chunk
     chunkNameExtractCss: '__KOOT__EXTRACT__CSS__',
@@ -63,7 +65,7 @@ module.exports = {
     filenameSPATemplateInjectJS: '.spa-template-inject.LOCALEID.js',
 
     // [生产环境] 文件名
-    filenameCurrentBundle: '.koot-current',
+    // filenameCurrentBundle: '.koot-current',
 
     // 标签属性
     styleTagGlobalAttributeName: 'data-koot-global',
@@ -85,5 +87,7 @@ module.exports = {
     thresholdScriptRunFirst: 15 * 1024,
 
     /** @type {string} 打包结果目录中的文件对照表文件名 */
-    buildManifestFilename: '.public-chunkmap.json'
+    buildManifestFilename: '.public-chunkmap.json',
+    /** @type {string} 打包结果目录中输出文件总集文件名 */
+    buildOutputsFilename: '.public-outputs.json'
 };